shellandtest: v2 - a better CLI library, zerolog, more functionality!
This commit has many changes. Firstly, the Go version has been bumped to
1.19.

One of the major changes is moving away from Cobra. While it is popular,
it has a lot of quirks to working with it. I picked the CLI library of Kong,
which is gaining a lot of traction after it's predecessor Kingpin was put into maintenance mode.

I picked this because it will be easier for someone to come in and help maintain this code
without having to follow all the caveats that I had to work around 3 years ago with Cobra.

The other major change is that I removed the ability to pipe a test file
into this application. Previously, it only sorta worked and in honesty,
I feel like being able to specify a directory or a single file with a
parameter will be better. I'm open to having my mind being changed. :)

Additionally, rather than just outputting things using fmt.Println, I
opted to utilize zerolog so that we can have logging with different log
levels, because in CI, we might only want "error", for example.

Furthermore, I ran this through a rather aggressive `golangci-lint`
configuration that I have and caught and fixed many errors, typos, and
added error checks for things that weren't checked previously.